// next/components/sections/Experience/types.ts
export interface ExperiencesTypes {
    company: string
    current: boolean | null
    description: string | null
    end_date: string | null
    id: string
    position?: string | null
    start_date: string
    technologies: string[] | null
    created_at: string | null
    updated_at?: string | null
}